move package location
