+++ pkg/webhook/validation/podspec_dryrun.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package extravalidation
+package validation
 
 import (
 	"context"
